// src/models/UsuarioQuiniela.ts
export interface UsuarioQuiniela {
    idProfile: string;
    participa1: boolean;
    participa2: boolean;
    participa3: boolean;
    participa4: boolean;
    participa5: boolean;
    participa6: boolean;
    partido1: number[];
    partido2: number[];
    partido3: number[];
    partido4: number[];
    partido5: number[];
    partido6: number[];
}